refactor(ci): skip release note reminder if PR explicitly says notes not needed
