+++ packages/@repo/release-notes/src/commands/createOrUpdateChangelogDocs.ts
@@ -19,11 +19,11 @@ import pMap from 'p-map'
 import {client} from '../client'
 import {STUDIO_PLATFORM_DOCUMENT_ID} from '../constants'
 import {type PullRequestInfo} from '../types'
-import {extractReleaseNotes} from '../utils/extractReleaseNotes'
 import {getCommits, getSemverTags} from '../utils/getCommits'
 import {getMergedPRForCommit} from '../utils/github'
 import {createId} from '../utils/ids'
 import {markdownToPortableText} from '../utils/portabletext-markdown/markdownToPortableText'
+import {extractReleaseNotes} from '../utils/pullRequestReleaseNotes'
 import {stripPr} from '../utils/stripPrNumber'
 import {uploadImages} from '../utils/uploadImages'
